cerebrum/example/agents/autogen_demo_agent/meta_requirements.txt
autogen-agentchat~=0.2